refactor(runner): add InferenceError to all pipelines
This commit adds the inference error logic from the SAM2 pipeline to all
pipelines so users are given a warning when they supply wrong arguments.
