+++ runner/app/routes/text_to_image.py
@@ -5,7 +5,8 @@ from typing import Annotated
 
 from app.dependencies import get_pipeline
 from app.pipelines.base import Pipeline
-from app.routes.util import HTTPError, ImageResponse, http_error, image_to_data_url
+from app.routes.utils import HTTPError, ImageResponse, http_error, image_to_data_url
+from app.utils.errors import InferenceError
 from fastapi import APIRouter, Depends, status
 from fastapi.responses import JSONResponse
 from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
@@ -142,19 +143,25 @@ async def text_to_image(
     has_nsfw_concept = []
     params.num_images_per_prompt = 1
     for seed in seeds:
+        params.seed = seed
+        kwargs = {k: v for k, v in params.model_dump().items() if k != "model_id"}
         try:
-            params.seed = seed
-            kwargs = {k: v for k, v in params.model_dump().items() if k != "model_id"}
             imgs, nsfw_check = pipeline(**kwargs)
-            images.extend(imgs)
-            has_nsfw_concept.extend(nsfw_check)
         except Exception as e:
             logger.error(f"TextToImagePipeline error: {e}")
             logger.exception(e)
+            if isinstance(e, InferenceError):
+                return JSONResponse(
+                    status_code=status.HTTP_400_BAD_REQUEST,
+                    content=http_error(str(e)),
+                )
+
             return JSONResponse(
                 status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
                 content=http_error("TextToImagePipeline error"),
             )
+        images.extend(imgs)
+        has_nsfw_concept.extend(nsfw_check)
 
     # TODO: Return None once Go codegen tool supports optional properties
     # OAPI 3.1 https://github.com/deepmap/oapi-codegen/issues/373
