+++ runner/app/routes/utils.py
@@ -70,22 +70,6 @@ class HTTPError(BaseModel):
     detail: APIError = Field(..., description="Detailed error information.")
 
 
-class InferenceError(Exception):
-    """Exception raised for errors during model inference."""
-
-    def __init__(self, message="Error during model execution", original_exception=None):
-        """Initialize the exception.
-
-        Args:
-            message: The error message.
-            original_exception: The original exception that caused the error.
-        """
-        if original_exception:
-            message = f"{message}: {original_exception}"
-        super().__init__(message)
-        self.original_exception = original_exception
-
-
 def http_error(msg: str) -> HTTPError:
     """Create an HTTP error response with the specified message.
